refactor all components
