+++ package.json
@@ -111,7 +111,7 @@
     "react-phone-input-2": "^2.9.4",
     "react-timekeeper": "^2.0.4",
     "react-tooltip": "^3.11.0",
-    "reactjs-popup": "^1.4.1",
+    "reactjs-popup": "^1.5.0",
     "styled-system": "^5.0.15"
   },
   "browserslist": {
