+++ src/components/expanding-search-input/styles.js
@@ -1,9 +1,5 @@
 import styled from 'styled-components'
-
-export const Container = styled.div`
-  display: flex;
-  justify-content: space-between;
-`
+import { radii, fontSizes } from '../../config/theme'
 
 export const ContainerAlt = styled.div`
   position: relative;
@@ -17,30 +13,44 @@ export const ContainerAlt = styled.div`
 export const Input = styled.input`
   border: none;
   background: transparent;
-  border-radius: 10em;
+  border-radius: ${({ theme }) => theme.radii.full};
   padding: 0.6em;
   margin-right: 1em;
   transition: transform 0.3s, opacity 0.3s;
   width: 15em;
-  transform: ${props => (props.expanded ? 'scaleX(1)' : 'scaleX(0)')};
+  transform: ${({expanded}) => expanded ? 'scaleX(1)' : 'scaleX(0)'};
   transform-origin: right;
-  opacity: ${props => (props.expanded ? '1' : '0')};
+  opacity: ${({expanded}) => expanded ? '1' : '0'};
   font-size: inherit;
 `
 
 export const InputAlt = styled.input`
   border: none;
-  background: #f0f0f0;
-  color: #999;
-  border-radius: 10em;
+  background: ${({ theme }) => theme.colors.gray.xlight};
+  color: ${({ theme }) => theme.colors.gray.dark};
+  border-radius: ${({ theme }) => theme.radii.xlarge};
   padding: 0.6em 1em;
   margin-right: 1em;
   transition: transform 0.3s, opacity 0.3s;
   width: 100%;
-  transform: ${props => (props.expanded ? 'scaleX(1)' : 'scaleX(0)')};
+  transform: ${({expanded}) => expanded ? 'scaleX(1)' : 'scaleX(0)'};
   transform-origin: right;
-  opacity: ${props => (props.expanded ? '1' : '0')};
-  font-size: ${props => props.theme.fontSizes.small};
-  font-family: inherit;
+  opacity: ${({expanded}) => expanded ? '1' : '0'};
+  font-size: ${({theme}) => theme.fontSizes.small};
+  font-family: ${({ theme }) => theme.fonts.Montserrat};
   height: 100%;
 `
+
+export const iconAltStyling = {
+  position: 'absolute',
+  right: 14,
+  borderRadius: radii.full,
+  padding: 9,
+  fontSize: fontSizes.xxlarge,
+  cursor: 'pointer',
+}
+
+export const expandedIconStyles = {
+  boxShadow: '0px 0px 16px 0px rgba(0, 0, 0, 0.09)',
+}
+
