+++ src/components/footer/index.js
@@ -1,23 +1,23 @@
 import * as React from 'react'
 import styled from 'styled-components'
 
-import Box from './box'
+import Box from '../box'
 
 const CopyrightText = styled.p.attrs({
   role: 'contentinfo',
 })`
   text-align: center;
   margin: 2em 0 0;
-  color: ${props => props.theme.colors.gray[3]};
-  font-size: ${props => props.theme.fontSizes.xsmall};
-  font-family: ${props => props.theme.fonts[0]};
-  font-weight: ${props => props.theme.fontWeights.bold};
+  color: ${({theme}) => theme.colors.gray.xdark};
+  font-size: ${({theme}) => theme.fontSizes.xsmall};
+  font-family: ${({theme}) => theme.fonts.Montserrat};
+  font-weight: ${({theme}) => theme.fontWeights.bold};
 `
 
 CopyrightText.displayName = 'CopyrightText'
 
-const Footer = ({ copyright, children, ...props }) => (
-  <Box {...props}>
+const Footer = ({ copyright, children, ...otherProps }) => (
+  <Box {...otherProps}>
     {children}
     <CopyrightText>{copyright}</CopyrightText>
   </Box>
