+++ src/components/theme-provider/index.js
@@ -2,7 +2,7 @@ import * as React from 'react'
 import PropTypes from 'prop-types'
 import { ThemeProvider as StyledThemeProvider } from 'styled-components'
 
-import baseTheme from './theme'
+import baseTheme from '../../config/theme'
 
 const ThemeProvider = ({ children }) => (
   <StyledThemeProvider theme={baseTheme}>{children}</StyledThemeProvider>
