+++ src/components/time-input/index.js
@@ -1,13 +1,12 @@
 import React, { Component } from 'react'
-import styled from 'styled-components'
 import TimeKeeper from 'react-timekeeper'
 import ClockOutline from '@lessondesk/material-icons/dist/ClockOutline'
 import { connect } from 'formik'
 
-import { colors } from '../theme'
+import { colors } from '../../config/theme'
 import './styles.css'
 
-import { errorForField } from '../utils'
+import { errorForField } from '../../utils/error-for-field'
 import InputWrapper from '../input-wrapper'
 import Input from '../styled-input'
