+++ src/utils/error-for-field.js
@@ -1,0 +1,6 @@
+export const errorForField = (errors, touched, fieldname) => {
+  if (errors[fieldname] && touched[fieldname]) {
+    return errors[fieldname] || ''
+  }
+  return ''
+}
